gg_file.cpp: #include "libgsft/gsft_strsep.h"
If the system doesn't provide strsep() in its C library, libgsft
will provide it.

Unlike the previous implementation of strtok_r(), the strsep() implementation
is not prefixed with gsft_. This is because gsft_strsep.c is only compiled
if the system does not provide strsep() itself. I will do something similar
later on for libgsft's copy of strdup().