libgsft: Added the BSD implementations of strlcpy() and strlcat().
gsft_strz.h has been renamed to gsft_szprintf.h, and it only contains
the szprintf() family of functions.

gsft_strlcpy.h contains definitions for strlcpy() and strlcat().

configure.ac has been updated to check for strlcpy() and strlcat()
in the system's C library, and if they don't exist, it enables the
internal versions.

TODO: Check the system version of strlcat() to make sure it functions
like the BSD version, and if it doesn't, enable the internal version.