[Win32] cc6.c: New function to check for Common Controls v6.0.
This function checks the version of COMCTL32.DLL instead of checking
the operating system version. A version of 5.82 or later indicates
Common Controls v6.0; 5.81 or earlier indicates an older version.

Wine 1.1.30's COMCTL32.DLL returns 5.81 regardless of whether or not
the operating system is set to Windows XP or later.