g_main_win32.cpp: Use AllocConsole() to allocate a console instead of a custom window.
There seems to be a bug in Wine 1.1.30 where appending enough text (~150 or
more characters) to a WC_EDIT control using Edit_SetSel() and Edit_ReplaceSel()
results in Wine freezing. So, use AllocConsole() to allocate a regular Win32
console window instead.

As an added bonus, this redirects stdout, stderr, and stdin to the console
window, which makes it easier to deal with.

The console code was adapted from
http://justcheckingonall.wordpress.com/2008/08/29/console-window-win32-app/