macros/log_msg.h: Added a macro LOG_MSG_ONCE().
This macro is similar to the Linux kernel's WARN_ONCE() macro. It logs a
message, and then sets a flag to prevent it from logging the same message
again.

This macro will be used for the upcoming runtime linking for zlib and
libpng. If the libraries aren't found, it will log a warning once, even
though it will attempt to reopen it every time it's requested.