src/gens/dll/dll_png.c: New dynamic loading infrastructure for libpng.
This will allow Gens/GS to be linked to libpng at runtime instead of compile-time.
The main benefit is that Gens/GS will still run on systems that don't have libpng,
but will "magically" gain PNG support if libpng is added later on.

On Windows, this means I can now distribute libpng in a separate DLL without
having to worry about the program not working if the DLL is missing.