configure.ac: Don't specify "-Wl," in GSFT_LDFLAGS_GCC_OPTION.
The extra comma confuses autoconf into creating two separate variables,
which resulted in the "--as-needed" detection result being used for the
"--hash-style=both" result. This broke on Win32, since Win32 only has one
hash style. (Either that, or it doesn't use hashes at all. I'm not sure.)

A workaround is to specify only the linker option as the parameter to the
macro and add -Wl to the LDFLAGS variable manually.