input_dinput.cpp: Removed usage of IDirectInputDevice2->GetCapabilities().
NeKit helped me to diagnose a bug that caused a crash in r7_pre4, but not
r7_pre3. The symptom was that loading a ROM via File, Open ROM resulted in
a crash, but loading a ROM via ROM History did not. git bisect reported
that the broken commit was 87bffefebd7e251a5f15022cdf1b701a95105468:
"input_dinput.cpp: Only set ranges of axes that exist on the joystick."

It turns out that NeKit has a Saitek P580 Color Rumble Pad [1], and the
driver for this pad doesn't like GetCapabilities(). It ends up corrupting
the heap, which only caused a problem when using File, Open ROM, since
that triggered input_dinput_set_cooperative_level_joysticks().

GetCapabilities() was used to determine if the X and Y axes were available,
since douglas previously reported that Gens/GS crashed when attempting to
set the axis ranges for those axes on his weird keyboard attachment, which
shows up as a joystick device. The axis ranges were being changed from
the DInput default of [0, 65535] to [-1000, 1000]. Interestingly enough,
the other axes were kept at the default setting. So, instead of trying to
figure out how to work around this driver bug, we're just going to leave
the axes at their default settings.

Additionally, some cleanups were done with regards to COM reference
counting and NULL pointers.

Special thanks to NeKit from Sonic Retro for helping to diagnose and
solve this problem.

[1] http://www.saitek.com/uk/prod/p580.htm