gens_ui_gtk.cpp::update(): Check the next process if the current process isn't finished.
This fixes a "freeze" bug I've noticed. I should have added this when I
created the wait list in the first place, but I forgot. :( (Effectively,
the bug made it no more useful than the previous method of waiting for
the process to finish before returning control from GensUI::LaunchBrowser().)