vdraw_text.cpp: Initial preparations for the VGA console font.
The VGA console font is 8x16. The text drawing code was hard-coded
to only support square characters. These preparations essentially
keep the same character size value, but double the height.

This is easier than adding support for completely different widths
and heights, though support for that may be added in the future if
I decide to allow the character set to be changed at runtime.