w32_unicode.c: Initial Win32 Unicode wrapper.
w32_unicode.c contains functions that convert UTF-8 to UTF-16 for Win32
Unicode support. Additionally, w32_unicode_init() creates function pointers
for either Win32 API or the wrapper functions, depending on if the system
supports Unicode or not.

Currently, the functions are hard-coded to Unicode (aka W functions).
This will be changed to dynamic loading of either A or W functions later on.

The Win32 Gens window and Color Adjustment window have had some initial work
on porting to the Unicode wrapper. This was done to check that the Unicode
wrapper was working properly. More windows and functions will be ported
later on.