[Win32] ccode_window.c: Ported the Country Code winodw to Unicode.
TODO: Figure out how to port LoadImage(). It accepts a "string", but
that string can either be a real string or a DWORD value crated by
MAKEINTRESOURCE() or MAKEINTRESOURCEW(). Then again, I'm not sure if
LoadImage() even needs to be ported to Unicode, since all it's used
for is loading images from inside the executable...

Added more macros and conversion functions to w32_unicode_x.c and
w32_unicode_x.h.