libgsft/w32u/w32u_priv.c::w32u_UTF8toUTF16(): Fixed a logic error that passed the wrong value for cchWideChar.
cchWideChar is the number of characters that can fit in lpWideCharStr.
w32u_UTF8toUTF16() was incorrectly passing the size of wcs in bytes.

I don't think this could have led to an overflow, since wcs was sized
correctly, but it's still possible that something could have tripped
it up. Plus, correct code is always better. :)