libgsft/w32u/w32u_commdlg.c: Ported this file to the new split-A/W system with compile-time linking.
The Open and Save functions have been combined into one function that accepts
a function pointer to the actual API function as an additional parameter.
The old functions still exist as wrappers. The new function is inlined in the
release build, so there's no extra overhead.

The ANSI version is not implemented yet.