input_sdl_events.cpp: If the vdraw backend is changed in windowed mode, restore the key modifiers.
Key modifiers are reset on backend change, since in full screen,
a backend change can take a few seconds. In windowed mode, backend
changes are nearly instantaneous, so the key modifiers should be
restored after the backend is changed.

The key modifiers are reset in vdraw_sdl_reinit_gens_window() and
vdraw_sdl_gl_reinit_gens_window(), which is why we have to restore
the modifiers afterwards instead of just saving them.