osd_font.c: Split the font loading functions from osd_charset.cpp into a separate file.
The font format has changed slightly. Instead of storing the character value
as a 32-bit value and flags as 16-bit, the character is stored as 24-bit and
the flags 8-bit. This works because Unicode defines U+10FFFF as the highest
character allowed.

I'm not increasing the file version, since no code has been released that
uses the old version.