mdp_host_gens_mem.c::mdp_host_mem_write_block_16(): Write to _32X_Rom if writing to MDP_MEM_MD_ROM and 32X is active.
This would potentially cause a problem if any plugins used this function
in 32X mode, since the 32X ROM (which is a copy of the ROM that's only
visible to the SH2s) isn't updated, but the regular ROM (which is visible
to the 68000 and Z80) is updated. However, no plugins included with Gens/GS
currently use this function, so it shouldn't be an issue.