ROM::loadROM(): If the ROM header couldn't be read, return early.
If no bytes were returned when reading the first 2 KB of the ROM header,
don't bother doing anything else.

Note that it is possible to have a ROM smaller than 2 KB, so we're not
checking to see if bytes_read matches 2 KB. Instead, we're just checking
to see if bytes_read is not zero.