gens_ui_gtk.cpp::UI_GTK_FileChooser(): If the initial file is a directory, set the current folder correctly.
This fixes an annoying bug where if you set the initial filename to a directory,
GTK+'s file chooser will start in the directory *below* the requested directory.
This is the correct behavior when choosing a new directory, but not for opening
or saving a file.

This fix calls gtk_file_chooser_set_current_folder() when the specified initial
file is a directory and we're opening or saving a file.