ui/win32/gens_ui_win32.cpp: Show a warning if the user is running as an administrator on NT-based systems.
Running as an administrator is bad, but people whine too much about
regular users not being administrator to disable administrator support
entirely. It may be disabled in a future release, though.

This check is only done on NT-based systems, since 9x doesn't support
permissions. (It's impossible to *not* run as an administrator on 9x.)