mdp/misc/vdp_debugger/: Initial VDP Debugger plugin for Linux/GTK+.
Currently, all it does is display the palette in realtime. Note that
the palette displayed is the palette stored in CRam after the frame
is finished, so raster effects (i.e. changing CRam after a scanline)
may result in the wrong palette being displayed.

Scanline events may be added to MDP 1.1 for improved accuracy.