vdp_debugger/vdpdbg_window_gtk2.c: Allow VDP registers to be edited by the user.
Double-clicking a register value will allow the user to edit the register by
entering a new value in decimal, hexadecimal (prefixed with 0x), or octal
(prefixed with 0).

Note that if the register is updated by the ROM while the user is editing it,
the edit will disappear. I'll figure out how to fix this later.