gens_core/vdp/vdp_rend_m5.cpp: Ported GET_PATTERN_DATA to C++.
This function gets a line of pattern data for the given tile for the current line.
Note that the original version took an extra parameter indicating window or scroll,
but that parameter was made obsolete a while ago and is no longer used, so I didn't
include it in the C++ version.

The original version returned two 32-bit parameters, pattern data and palette
number. The new version only returns the pattern data. The palette number is
extracted by the caller using the original pattern info.

TODO: Write a C++ inline function to extract the palette number from pattern info.