gens_core/vdp/vdp_rend_m5.cpp: Ported UPDATE_MASK_SPRITE to C++.
This was probably the hardest macro to port, since it kept having issues
at the beginning. Turns out that it was because RENDER_LINE_SPR expected
UPDATE_MASK_SPRITE to set %edx to VDP_Current_Line, even though this
wasn't mentioned in the comments for either macro.

As an extra note, this port seems to fix Test 6 in Nemesis' Sprite Masking
and Overflow test ROM. Previously, Test 6 would say "Pass", but the bottom
row of said "Pass" was red, which indicates a partial failure. Now, the
bottom row is green. :)