gens_core/vdp/vdp_rend_m5.cpp: Ported RENDER_LINE_SCROLL_B to C++.
We're almost finished porting the MD VDP code to C++! :) We just need to port
the Scroll A / Window renderer, the Sprite renderer, and the rest of Render_Line().
Then we get to do some fun stuff, like putting border color emulation right
in the VDP rendering code instead of in the VDraw backends.