gens_core/vdp/vdp_io.h: Renamed vdp_isV30() to vdp_getVPix().
This is the same sort of update that was done for vdp_getHPix().
It now returns either 240 or 224, depending on the VDP mode and if
a game is loaded and/or the debugger is active.

This may also return 192 in the future, once SMS support has been added.
(144 is unlikely, since the Game Gear does actually render a standard
SMS-sized buffer. It simply displays a small portion of it.)