debugger/debugger.cpp::T_Refresh_VDP_Palette_Colors(): Make sure the palettes are updated.
If a savestate is loaded while the debugger is running, the CRAM will be
updated with the contents of the savestate, but the palette won't be updated,
since the main game loop isn't running. Check VDP_Flags.CRam in the VDP
debugger, and if it needs to be updated, update it.