debugger/debugger.cpp: Update the palette before the patterns.
If the palette isn't updated first, then if a savestate is loaded while
the debugger is active, the new patterns will be loaded with the old
palette first, *then* they will be updated with the new palette.

Updating the palette before the patterns  ensures that the patterns are
drawn with the correct palette in the first place.