debugger/debugger.cpp: Improved line number display for SegaCD Word RAM patterns.
Word RAM patterns can be arranged in 16x16 or 32x32 patterns. The line numbers
were listed in the same fashion as the regular Genesis VDP, which is 8x8 only.
So, we should only display line numbers for every other line in 16x16, or every
four lines in 32x32.

TODO: The MD VDP actually supports 8x16 in interlaced mode. Add a viewer for this.