gens_core/vpd/vdp_io.h: Renamed the anonymous register struct in VDP_Reg to "m5".
These registers are named for Mode 5 and may be different for other modes.

Added comments for each register in VDP_Reg.m5 to indicate the function of
the register, including bit descriptions for mode setting registers, the
scroll size register, and the interlace register.

Updated all code that uses VDP_Reg to reference VDP_Reg.m5.