gens_core/vdp/vdp_io.c: Ported Write_Byte_VDP_Data() to C [renamed to VDP_Write_Data_Byte()].
This function is merely a wrapper function for Write_Word_VDP_Data(). On the
actual hardware, 8-bit writes to the VDP port result in the byte being mirrored
for both the high byte and low byte of the 16-bit word, so this function does
that mirroring and then calls Write_Byte_VDP_Data().