gens_core/vdp/vdp_io.c::VDP_Update_DMA(): Fixed the DMA timing check.
VDP_Reg.DMAT_Length is unsigned, so checking if it's less than zero always
returns false. Instead, check if VDP_Reg.DMAT_Length is greater than timing.

The subtraction was an optimization in the assembler version; however, it's
not needed in the C version, since the subtraction is only needed if the DMA
isn't finished. If it is finished, DMAT_Length is reset to 0.