gens_core/vdp/vdp_io.c: New VDP_Lines_t struct for VDP line information.
This struct contains all the information about the entire display size
and the visible size. Among other things, this allows us to draw the top
border in the correct frame. (Previously, the top border was drawn from
the previous frame, which is wrong.)

Updated g_md.cpp::T_gens_do_MD_frame() to use a single loop for the entire
frame. g_mcd.cpp and g_32x.cpp are currently #if 0'd.