emulator/g_32x.cpp: Updated T_gens_do_32X_frame() to support VDP_Lines.
It doesn't properly account for the border yet.

Updated gens_core/vdp/vdp_rend_x86.asm to ensure that the current line
is in range (0-239) before attempting to render anything. (Note that in
V28 mode, the full image is drawn anyway, which can lead to screwy stuff
showing up in the bottom 16 pixels of the screen.)