vdp_io.c: Ported the DMA length check from asm to C++.
I removed the "VRam Only" part of the check, since that was actually a
different issue: DMA to CRam/VSRam terminates if the address counter
is >=0x80. I didn't know this when I originally implemented the zero-length
DMA check, so Gens/GS never checked for it.

(As a side note, the DMA_LOOP code bitwise-AND's the address with 0x7E when
doing DMA to CRam/VSRam. It should simply terminate once it hits >=0x80.)