vdp_io.cpp: Initial port of DMA_LOOP to C++.
Renamed vdp_io.c to vdp_io.cpp.

The initial port of DMA_LOOP (named T_DMA_Loop()) only does RAM to VRam.

I removed part of the end of the DMA code that was never used. Specifically,
the part that marks DMA as finished if there's "nothing to do". This section
was never called! (I added a divide-by-zero there to trap cases in which it
was called, and it never happened.)