gens_core/vdp/vdp_io.cpp::T_DMA_Loop(): Increment dest_address using VDP_Reg.m5.Auto_Inc.
This fixes the intro animation in Sonic 3D, which alternates the auto-increment
register between +4 and 0. T_DMA_Loop() was assuming that auto-increment was
always set to +2. The original x86 asm version correctly used the auto-increment
register, so this is a regression in the C++ port.

I also did a few other fixes in T_DMA_Loop() that should hopefully fix cases
that I haven't been able to test, including an incorrectly-shifted address
for SegaCD cell conversion and correct address increments for the DMA source
address. I also removed the switch/case fallthrough for CRam and VSRam, since
that made it more difficult to understand exactly what was happening.