video/vdraw_ddraw.cpp: Updated to work with horizontal border drawing.
NOTE: This commit removes the code that maintained the aspect ratio of the
image regardless of the window size. I probably won't be readding it, since
DirectDraw is a pain to work with. Instead, I'll figure out how to get the
OpenGL projection code working and port the OpenGL renderer to Win32.