vdp_io.cpp: Check interlaced mode at the start of the frame.
Changes to the interlaced mode setting only take effect at the start
of the next frame. Set VDP_Reg.Interlaced (a new field) at the start
of the frame if it's interlaced mode 2, and clear it if it's not.

T_Make_Sprite_Struct(): Use (VDP_Reg.H_Cell * 2) to calculate the maximum
number of sprites onscreen instead of testing bits in the VDP register.