vdp_io.cpp::DMA_Fill(): Fixed a weird DMA Fill issue.
The original asm code looked wrong compared to the official documentation,
so I changed it. Turns out the asm code was right. The change caused weird
dots to appear in the title cards in Sonic 2's 2P VS mode if Odd Fields Only
or Flickering Interlaced were selected.

Switching back to the method used in the original asm version fixes this problem.