g_md.cpp, g_mcd.cpp, g_32x.cpp: Clear the VBlank status flag in the visible area.
This fixes Gauntlet 4. Apparently, Gauntlet 4 only checks the start button
on the title screen if the VBlank flag is set for enough lines. It worked
with the old line numbering system, since display line 0 was the same as
visible line 0, but it broke horribly with the new line numbering system,
since the VBlank status flag was being reset at a negative visible line.
(I have no idea why Gauntlet 4 does this, but whatever.)

TODO: Only clear the VBlank status flag on visible line 0.