vdp_32x.c: Combined _32X_Palette_16B[] and _32X_Palette_32B[] into a union.
This will save around 64 KB of memory.

Updated the rest of Gens/GS to work with the new unionized _32X_Palette.

md_palette.cpp: Added a new function Adjust_CRam_32X() that adjusts the 32X CRam.
This function is actually a wrapper around a new templated function, which is used
in Recalculate_Palettes(). Adjust_CRam_32X() is used in save.cpp::GsxImport32X().