vdp_rend_m5.cpp::T_Render_LineBuf_32X(): Optimized many of the rendering methods.
I timed Render Mode 5, which is used in Knuckles' Chaotix for the main game,
by running the rendering function 1,000,000 times and taking the time difference
between the start of the iterations and the end of the iterations.

1,000,000 iterations for render mode 5 (chaotix main game)

          Old    New    Diff      %
Debug:   3.44s  2.43s  -1.01s  -29.4%
Release: 1.38s  0.78s  -0.60s  -43.5%

That's a 43.5% performance increase in the release build! :D