vdp_rend_m5.cpp::T_Render_LineBuf_32X(): Don't divide VDP_Reg.H_Pix when initializing px.
Instead, decrement px by 4 or 2. This eliminates an unnecessary shift.

The reason why this was done before is because 320 used to be hard-coded
as the 32X horizontal resolution, so gcc optimized it out to the correct
constant value. This was changed to use a variable in order to support
H32 mode, so the extra shift isn't needed.