libgsft/w32u/w32u_winnls.c: Implemented an emulation function for GetCPInfoEx().
Windows NT 4.0 and Windows 95 do not have the GetCPInfoExA() or GetCPInfoExW()
functions, so Gens/GS refused to start up due to an unresolved symbol.

The emulation function provides a mostly-complete implementation of the
GetCPInfoEx() function. It calls GetCPInfo() first in order to fill in the
first three fields, then fills in the CodePage and UnicodeDefaultChar.
CodePageName is blanked for now, since there's no way to resolve a code page
number to a name on Windows without using the real GetCPInfoEx() function.