debugger/fb_32x.cpp: Draw the 32X CRam at the bottom of the screen.
TODO: Draw the entire visible area for the 32X framebuffer, and move the
palette to the border region. This may cause an overlay issue if a game
uses V30, but since the majority of games use V28, it shouldn't be that
big of a problem.

TODO: Maybe move SegaCD LEDs to the border region, too?