emulator/g_main_win32.cpp::get_default_save_path(): Fixed support for Win9x and WinNT 4.0.
On Windows 9x, user files are stored in the Windows directory (e.g. C:\Windows\Desktop).
Thus, we can't check the Windows directory on those operating systems.

On Windows NT 4.0, user profiles are stored in the Windows directory
(e.g. C:\WinNT\Profiles). So we have to check if the user profile
is part of the pathname. If it is, then we use the current directory.