vdp_io.cpp::VDP_Read_Data(): Attempt to work around a gcc optimization bug.
This bug seems to cause text corruption in Soleil (Crusader of Centry)
when Gens/GS is compiled with optimizations. Marking the data variable
as volatile seems to fix it for both gcc-4.4 (Linux) and gcc-4.3 (mingw).

TODO: Report this as a bug to the gcc developers.

Related bug report: http://gs_server.gerbilsoft.ddns.info/bugs/show_bug.cgi?id=33