mem_m68k.asm: Fixed some regressions caused by the VDP mirroring changes.
This is a regression from commit 3c9f57577ff5f8d5a85e6931f2fd330db986f47f.

I updated the VDP memory functions to check for all VDP mirrors. If the
address matched a mirror, then the address was AND'd with 0x1F to get
the actual VDP address. However, I forgot to update some of the internal
VDP address checks to use the 5-bit addresses, so it ended up causing
random breakage.

This bug was reported by Jorge in #retro.