vdp_io.cpp::VDP_Update_DMA(): Fixed a timing regression.
This fixes a regression from commit 967f13f88a63cb3e01ae730356b4b64d7ed3f3ea.

I noticed in Jorge's Sonic 2 hack that the bar between the splitscreen in
2P VS was one line larger in r7+ than in r7. I verified the same issue in
regular Sonic 2.

Turns out this was a regression from a small change in how timing was
calculated. The code was originally subtracting "timing" from VDP_Reg.DMAT_Length
in order to check if the DMA was finished. I removed this and did a simple
comparison instead; however, I forgot to remove the code that re-added
"timing" to VDP_Reg.DMAT_Length to compensate for the subtraction in the
case that the DMA was finished.